Edited appy.c
Sorry for such an extensive revision...
Reformated comments to be blocks instead of lines.
Changed from raw to cbreak input.
Added signal processing for SIGINT.
Removed homebrew CTRL-C detection.
Reformated TODO comments to be more readable, and included general todo in
it's own section at the bottom of the source file.
Added ncurses extension headers, and separated ncurses headers from
standard headers.
Created a function to contain memory garbage collection, and to be the
single exit point for the application.
Made better use of vertical space, for readability.
Added an application state and function to begin drawing the UI.